// bot/types/Opportunity.ts
export interface Opportunity {

    ////////////////////////////////////////////////////////////
    // SOURCE
    ////////////////////////////////////////////////////////////

    buyDex: string;

    sellDex: string;

    buyAdapter: string;

    sellAdapter: string;

    ////////////////////////////////////////////////////////////
    // POOL
    ////////////////////////////////////////////////////////////

    buyPool: string;

    sellPool: string;

    ////////////////////////////////////////////////////////////
    // TOKEN
    ////////////////////////////////////////////////////////////

    tokenIn: string;

    tokenOut: string;

    ////////////////////////////////////////////////////////////
    // UNISWAP
    ////////////////////////////////////////////////////////////

    buyFee?: number;

    sellFee?: number;

    ////////////////////////////////////////////////////////////
    // AERODROME
    ////////////////////////////////////////////////////////////

    buyStable?: boolean;

    sellStable?: boolean;

    ////////////////////////////////////////////////////////////
    // AMOUNTS
    ////////////////////////////////////////////////////////////

    loanAmount: bigint;

    amountOut: bigint;

    amountBack: bigint;

    ////////////////////////////////////////////////////////////
    // USD VALUES
    ////////////////////////////////////////////////////////////

    loanAmountUSD: number;

    grossProfitUSD: number;

    gasCostUSD?: number;

    flashLoanFeeUSD?: number;

    netProfitUSD?: number;

    ////////////////////////////////////////////////////////////
    // GAS
    ////////////////////////////////////////////////////////////

    gasLimit?: bigint;

    gasPrice?: bigint;

    gasRatio?: number;

    ////////////////////////////////////////////////////////////
    // EXECUTION
    ////////////////////////////////////////////////////////////

    profitable?: boolean;

    reason?: string;

    ////////////////////////////////////////////////////////////
    // DEBUG
    ////////////////////////////////////////////////////////////

    buyData?: string;

    sellData?: string;

    minProfit?: bigint;

    timestamp?: number;

}